♻️ カスタムフックにロジックを分割
